+++ src/base.rs
@@ -48,17 +48,10 @@ macro_rules! impl_noise {
             }
         }
 
-        impl $crate::Sample<2> for $crate::Seeded<$struct> {
+        impl $crate::SampleWithSeed<2> for $struct {
             #[inline(always)]
-            fn sample(&self, point: [f32; 2]) -> f32 {
-                self.noise.gen2(point, self.seed)
-            }
-        }
-
-        impl $crate::Sample<2> for $crate::Seeded<&$struct> {
-            #[inline(always)]
-            fn sample(&self, point: [f32; 2]) -> f32 {
-                self.noise.gen2(point, self.seed)
+            fn sample_with_seed(&self, point: [f32; 2], seed: i32) -> f32 {
+                self.gen2(point, seed)
             }
         }
 
@@ -71,18 +64,10 @@ macro_rules! impl_noise {
         }
 
         #[cfg(feature = "nightly-simd")]
-        impl $crate::Sample<2, core::simd::f32x2> for $crate::Seeded<$struct> {
+        impl $crate::SampleWithSeed<2, core::simd::f32x2> for $struct {
             #[inline(always)]
-            fn sample(&self, point: core::simd::f32x2) -> f32 {
-                self.noise.gen2a(point, self.seed)
-            }
-        }
-
-        #[cfg(feature = "nightly-simd")]
-        impl $crate::Sample<2, core::simd::f32x2> for $crate::Seeded<&$struct> {
-            #[inline(always)]
-            fn sample(&self, point: core::simd::f32x2) -> f32 {
-                self.noise.gen2a(point, self.seed)
+            fn sample_with_seed(&self, point: core::simd::f32x2, seed: i32) -> f32 {
+                self.gen2a(point, seed)
             }
         }
 
@@ -94,17 +79,10 @@ macro_rules! impl_noise {
                 }
             }
 
-            impl $crate::Sample<3> for $crate::Seeded<$struct> {
+            impl $crate::SampleWithSeed<3> for $struct {
                 #[inline(always)]
-                fn sample(&self, point: [f32; 3]) -> f32 {
-                    self.noise.gen3(point, self.seed)
-                }
-            }
-
-            impl $crate::Sample<3> for $crate::Seeded<&$struct> {
-                #[inline(always)]
-                fn sample(&self, point: [f32; 3]) -> f32 {
-                    self.noise.gen3(point, self.seed)
+                fn sample_with_seed(&self, point: [f32; 3], seed: i32) -> f32 {
+                    self.gen3(point, seed)
                 }
             }
 
@@ -117,18 +95,10 @@ macro_rules! impl_noise {
             }
 
             #[cfg(feature = "nightly-simd")]
-            impl $crate::Sample<3, core::simd::f32x4> for $crate::Seeded<$struct> {
+            impl $crate::SampleWithSeed<3, core::simd::f32x4> for $struct {
                 #[inline(always)]
-                fn sample(&self, point: core::simd::f32x4) -> f32 {
-                    self.noise.gen3a(point, self.seed)
-                }
-            }
-
-            #[cfg(feature = "nightly-simd")]
-            impl $crate::Sample<3, core::simd::f32x4> for $crate::Seeded<&$struct> {
-                #[inline(always)]
-                fn sample(&self, point: core::simd::f32x4) -> f32 {
-                    self.noise.gen3a(point, self.seed)
+                fn sample_with_seed(&self, point: core::simd::f32x4, seed: i32) -> f32 {
+                    self.gen3a(point, seed)
                 }
             }
         }
@@ -141,17 +111,10 @@ macro_rules! impl_noise {
                 }
             }
 
-            impl $crate::Sample<4> for $crate::Seeded<$struct> {
+            impl $crate::SampleWithSeed<4> for $struct {
                 #[inline(always)]
-                fn sample(&self, point: [f32; 4]) -> f32 {
-                    self.noise.gen4(point, self.seed)
-                }
-            }
-
-            impl $crate::Sample<4> for $crate::Seeded<&$struct> {
-                #[inline(always)]
-                fn sample(&self, point: [f32; 4]) -> f32 {
-                    self.noise.gen4(point, self.seed)
+                fn sample_with_seed(&self, point: [f32; 4], seed: i32) -> f32 {
+                    self.gen4(point, seed)
                 }
             }
 
@@ -164,18 +127,10 @@ macro_rules! impl_noise {
             }
 
             #[cfg(feature = "nightly-simd")]
-            impl $crate::Sample<4, core::simd::f32x4> for $crate::Seeded<$struct> {
+            impl $crate::SampleWithSeed<4, core::simd::f32x4> for $struct {
                 #[inline(always)]
-                fn sample(&self, point: core::simd::f32x4) -> f32 {
-                    self.noise.gen4a(point, self.seed)
-                }
-            }
-
-            #[cfg(feature = "nightly-simd")]
-            impl $crate::Sample<4, core::simd::f32x4> for $crate::Seeded<&$struct> {
-                #[inline(always)]
-                fn sample(&self, point: core::simd::f32x4) -> f32 {
-                    self.noise.gen4a(point, self.seed)
+                fn sample_with_seed(&self, point: core::simd::f32x4, seed: i32) -> f32 {
+                    self.gen4a(point, seed)
                 }
             }
         }