+++ src/fractal/fbm.rs
@@ -3,7 +3,7 @@ use core::simd::{LaneCount, Simd, SupportedLaneCount};
 
 use crate::{
     math::{fast_min, lerp},
-    Frequency, Sample, Seeded,
+    Frequency, Sample, SampleWithSeed, Seeded,
 };
 
 #[cfg(feature = "nightly-simd")]
@@ -53,7 +53,7 @@ impl<Noise> Fbm<Noise> {
 
 impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Fbm<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
     fn sample(&self, point: [f32; DIM]) -> f32 {
@@ -68,21 +68,18 @@ where
     }
 }
 
-impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Seeded<Fbm<Noise>>
+impl<Noise, const DIM: usize> SampleWithSeed<DIM, [f32; DIM]> for Fbm<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
-    fn sample(&self, point: [f32; DIM]) -> f32 {
-        let &Seeded {
-            noise: Fbm {
-                ref noise,
-                octaves,
-                gain,
-                lacunarity,
-                fractal_bounding,
-            },
-            seed,
+    fn sample_with_seed(&self, point: [f32; DIM], seed: i32) -> f32 {
+        let &Fbm {
+            ref noise,
+            octaves,
+            gain,
+            lacunarity,
+            fractal_bounding,
         } = self;
         fbm(noise, octaves, gain, lacunarity, fractal_bounding, seed, point)
     }
@@ -90,7 +87,7 @@ where
 
 impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Weighted<Fbm<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
     fn sample(&self, point: [f32; DIM]) -> f32 {
@@ -109,26 +106,21 @@ where
     }
 }
 
-impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Seeded<Weighted<Fbm<Noise>>>
+impl<Noise, const DIM: usize> SampleWithSeed<DIM, [f32; DIM]> for Weighted<Fbm<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
-    fn sample(&self, point: [f32; DIM]) -> f32 {
-        let &Seeded {
-            noise:
-                Weighted {
-                    fractal:
-                        Fbm {
-                            ref noise,
-                            octaves,
-                            gain,
-                            lacunarity,
-                            fractal_bounding,
-                        },
-                    strength: weighted_strength,
-                },
-            seed,
+    fn sample_with_seed(&self, point: [f32; DIM], seed: i32) -> f32 {
+        let &Weighted {
+            fractal: Fbm {
+                ref noise,
+                octaves,
+                gain,
+                lacunarity,
+                fractal_bounding,
+            },
+            strength: weighted_strength,
         } = self;
         weighted_fbm(noise, octaves, gain, lacunarity, fractal_bounding, weighted_strength, seed, point)
     }
@@ -137,7 +129,7 @@ where
 #[cfg(feature = "nightly-simd")]
 impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Fbm<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
@@ -154,22 +146,19 @@ where
 }
 
 #[cfg(feature = "nightly-simd")]
-impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Seeded<Fbm<Noise>>
+impl<Noise, const DIM: usize, const LANES: usize> SampleWithSeed<DIM, Simd<f32, LANES>> for Fbm<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
-    fn sample(&self, point: Simd<f32, LANES>) -> f32 {
-        let &Seeded {
-            noise: Fbm {
-                ref noise,
-                octaves,
-                gain,
-                lacunarity,
-                fractal_bounding,
-            },
-            seed,
+    fn sample_with_seed(&self, point: Simd<f32, LANES>, seed: i32) -> f32 {
+        let &Fbm {
+            ref noise,
+            octaves,
+            gain,
+            lacunarity,
+            fractal_bounding,
         } = self;
         fbm_a(noise, octaves, gain, lacunarity, fractal_bounding, seed, point)
     }
@@ -178,7 +167,7 @@ where
 #[cfg(feature = "nightly-simd")]
 impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Weighted<Fbm<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
@@ -199,27 +188,22 @@ where
 }
 
 #[cfg(feature = "nightly-simd")]
-impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Seeded<Weighted<Fbm<Noise>>>
+impl<Noise, const DIM: usize, const LANES: usize> SampleWithSeed<DIM, Simd<f32, LANES>> for Weighted<Fbm<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
-    fn sample(&self, point: Simd<f32, LANES>) -> f32 {
-        let &Seeded {
-            noise:
-                Weighted {
-                    fractal:
-                        Fbm {
-                            ref noise,
-                            octaves,
-                            gain,
-                            lacunarity,
-                            fractal_bounding,
-                        },
-                    strength: weighted_strength,
-                },
-            seed,
+    fn sample_with_seed(&self, point: Simd<f32, LANES>, seed: i32) -> f32 {
+        let &Weighted {
+            fractal: Fbm {
+                ref noise,
+                octaves,
+                gain,
+                lacunarity,
+                fractal_bounding,
+            },
+            strength: weighted_strength,
         } = self;
 
         weighted_fbm_a(noise, octaves, gain, lacunarity, fractal_bounding, weighted_strength, seed, point)
@@ -229,13 +213,13 @@ where
 #[inline(always)]
 fn fbm<Noise, const DIM: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, mut seed: i32, mut point: [f32; DIM]) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = Seeded { noise, seed }.sample(point);
+        let noise = noise.sample_with_seed(point, seed);
         seed = seed.wrapping_add(1);
         sum += noise * amp;
 
@@ -253,14 +237,14 @@ where
 #[inline(always)]
 fn fbm_a<Noise, const DIM: usize, const LANES: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, mut seed: i32, mut point: Simd<f32, LANES>) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = Seeded { noise, seed }.sample(point);
+        let noise = noise.sample_with_seed(point, seed);
         seed = seed.wrapping_add(1);
         sum += noise * amp;
 
@@ -274,13 +258,13 @@ where
 #[inline(always)]
 fn weighted_fbm<Noise, const DIM: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, weighted_strength: f32, mut seed: i32, mut point: [f32; DIM]) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = Seeded { noise, seed }.sample(point);
+        let noise = noise.sample_with_seed(point, seed);
         seed = seed.wrapping_add(1);
         sum += noise * amp;
         amp *= lerp(1.0, fast_min(noise + 1.0, 2.0) * 0.5, weighted_strength);
@@ -308,14 +292,14 @@ fn weighted_fbm_a<Noise, const DIM: usize, const LANES: usize>(
     mut point: Simd<f32, LANES>,
 ) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = Seeded { noise, seed }.sample(point);
+        let noise = noise.sample_with_seed(point, seed);
         seed = seed.wrapping_add(1);
         sum += noise * amp;
         amp *= lerp(1.0, fast_min(noise + 1.0, 2.0) * 0.5, weighted_strength);