+++ src/fractal/ridged.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "nightly-simd")]
 use core::simd::{LaneCount, Simd, SupportedLaneCount};
 
-use crate::{math::lerp, Frequency, Sample, Seeded};
+use crate::{math::lerp, Frequency, Sample, SampleWithSeed, Seeded};
 
 #[cfg(feature = "nightly-simd")]
 use crate::math::splat;
@@ -47,7 +47,7 @@ impl<Noise> Ridged<Noise> {
 
 impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Ridged<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
     fn sample(&self, point: [f32; DIM]) -> f32 {
@@ -64,22 +64,19 @@ where
     }
 }
 
-impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Seeded<Ridged<Noise>>
+impl<Noise, const DIM: usize> SampleWithSeed<DIM, [f32; DIM]> for Ridged<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
-    fn sample(&self, point: [f32; DIM]) -> f32 {
-        let &Seeded {
-            noise: Ridged {
-                ref noise,
-                octaves,
-                gain,
-                lacunarity,
-                fractal_bounding,
-                ..
-            },
-            seed,
+    fn sample_with_seed(&self, point: [f32; DIM], seed: i32) -> f32 {
+        let &Ridged {
+            ref noise,
+            octaves,
+            gain,
+            lacunarity,
+            fractal_bounding,
+            ..
         } = self;
 
         ridged(noise, octaves, gain, lacunarity, fractal_bounding, seed, point)
@@ -88,7 +85,7 @@ where
 
 impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Weighted<Ridged<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
     fn sample(&self, point: [f32; DIM]) -> f32 {
@@ -107,26 +104,21 @@ where
     }
 }
 
-impl<Noise, const DIM: usize> Sample<DIM, [f32; DIM]> for Seeded<Weighted<Ridged<Noise>>>
+impl<Noise, const DIM: usize> SampleWithSeed<DIM, [f32; DIM]> for Weighted<Ridged<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     #[inline]
-    fn sample(&self, point: [f32; DIM]) -> f32 {
-        let &Seeded {
-            noise:
-                Weighted {
-                    fractal:
-                        Ridged {
-                            ref noise,
-                            octaves,
-                            gain,
-                            lacunarity,
-                            fractal_bounding,
-                        },
-                    strength: weighted_strength,
-                },
-            seed,
+    fn sample_with_seed(&self, point: [f32; DIM], seed: i32) -> f32 {
+        let &Weighted {
+            fractal: Ridged {
+                ref noise,
+                octaves,
+                gain,
+                lacunarity,
+                fractal_bounding,
+            },
+            strength: weighted_strength,
         } = self;
 
         weighted_ridged(noise, octaves, gain, lacunarity, fractal_bounding, weighted_strength, seed, point)
@@ -136,11 +128,23 @@ where
 #[cfg(feature = "nightly-simd")]
 impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Ridged<Noise>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
     fn sample(&self, point: Simd<f32, LANES>) -> f32 {
+        self.sample_with_seed(point, 0)
+    }
+}
+
+#[cfg(feature = "nightly-simd")]
+impl<Noise, const DIM: usize, const LANES: usize> SampleWithSeed<DIM, Simd<f32, LANES>> for Ridged<Noise>
+where
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
+    LaneCount<LANES>: SupportedLaneCount,
+{
+    #[inline]
+    fn sample_with_seed(&self, point: Simd<f32, LANES>, seed: i32) -> f32 {
         let &Ridged {
             ref noise,
             octaves,
@@ -150,30 +154,6 @@ where
             ..
         } = self;
 
-        ridged_a(noise, octaves, gain, lacunarity, fractal_bounding, 0, point)
-    }
-}
-
-#[cfg(feature = "nightly-simd")]
-impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Seeded<Ridged<Noise>>
-where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
-    LaneCount<LANES>: SupportedLaneCount,
-{
-    #[inline]
-    fn sample(&self, point: Simd<f32, LANES>) -> f32 {
-        let &Seeded {
-            noise: Ridged {
-                ref noise,
-                octaves,
-                gain,
-                lacunarity,
-                fractal_bounding,
-                ..
-            },
-            seed,
-        } = self;
-
         ridged_a(noise, octaves, gain, lacunarity, fractal_bounding, seed, point)
     }
 }
@@ -181,7 +161,7 @@ where
 #[cfg(feature = "nightly-simd")]
 impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Weighted<Ridged<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
@@ -202,27 +182,22 @@ where
 }
 
 #[cfg(feature = "nightly-simd")]
-impl<Noise, const DIM: usize, const LANES: usize> Sample<DIM, Simd<f32, LANES>> for Seeded<Weighted<Ridged<Noise>>>
+impl<Noise, const DIM: usize, const LANES: usize> SampleWithSeed<DIM, Simd<f32, LANES>> for Weighted<Ridged<Noise>>
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     #[inline]
-    fn sample(&self, point: Simd<f32, LANES>) -> f32 {
-        let &Seeded {
-            noise:
-                Weighted {
-                    fractal:
-                        Ridged {
-                            ref noise,
-                            octaves,
-                            gain,
-                            lacunarity,
-                            fractal_bounding,
-                        },
-                    strength: weighted_strength,
-                },
-            seed,
+    fn sample_with_seed(&self, point: Simd<f32, LANES>, seed: i32) -> f32 {
+        let &Weighted {
+            fractal: Ridged {
+                ref noise,
+                octaves,
+                gain,
+                lacunarity,
+                fractal_bounding,
+            },
+            strength: weighted_strength,
         } = self;
 
         weighted_ridged_a(noise, octaves, gain, lacunarity, fractal_bounding, weighted_strength, seed, point)
@@ -232,13 +207,13 @@ where
 #[inline(always)]
 fn ridged<Noise, const DIM: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, mut seed: i32, mut point: [f32; DIM]) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = fast_abs(Seeded { noise, seed }.sample(point));
+        let noise = fast_abs(noise.sample_with_seed(point, seed));
         seed = seed.wrapping_add(1);
         sum += (noise * -2.0 + 1.0) * amp;
 
@@ -256,14 +231,14 @@ where
 #[inline(always)]
 fn ridged_a<Noise, const DIM: usize, const LANES: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, mut seed: i32, mut point: Simd<f32, LANES>) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = Seeded { noise, seed }.sample(point);
+        let noise = noise.sample_with_seed(point, seed);
         seed = seed.wrapping_add(1);
         sum += (noise * -2.0 + 1.0) * amp;
 
@@ -277,13 +252,13 @@ where
 #[inline(always)]
 fn weighted_ridged<Noise, const DIM: usize>(noise: &Noise, octaves: u32, gain: f32, lacunarity: f32, fractal_bounding: f32, weighted_strength: f32, mut seed: i32, mut point: [f32; DIM]) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, [f32; DIM]>,
+    Noise: SampleWithSeed<DIM, [f32; DIM]>,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = fast_abs(Seeded { noise, seed }.sample(point));
+        let noise = fast_abs(noise.sample_with_seed(point, seed));
         seed = seed.wrapping_add(1);
         sum += (noise * -2.0 + 1.0) * amp;
         amp *= lerp(1.0, 1.0 - noise, weighted_strength);
@@ -311,14 +286,14 @@ fn weighted_ridged_a<Noise, const DIM: usize, const LANES: usize>(
     mut point: Simd<f32, LANES>,
 ) -> f32
 where
-    for<'a> Seeded<&'a Noise>: Sample<DIM, Simd<f32, LANES>>,
+    Noise: SampleWithSeed<DIM, Simd<f32, LANES>>,
     LaneCount<LANES>: SupportedLaneCount,
 {
     let mut sum = 0.0;
     let mut amp = fractal_bounding;
 
     for _ in 0..octaves {
-        let noise = fast_abs(Seeded { noise, seed }.sample(point));
+        let noise = fast_abs(noise.sample_with_seed(point, seed));
         seed = seed.wrapping_add(1);
         sum += (noise * -2.0 + 1.0) * amp;
         amp *= lerp(1.0, 1.0 - noise, weighted_strength);