+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - **breaking:** `CellValue`, `CellDistance` and `CellDistanceSq` now have the `Fast*` prefix. They also have a `jitter` field. There is no more `Jitter` wrapper struct.
+- **breaking:** sampling with a seed is now expressed by the trait `SampleWithSeed` instead of `Seeded<T>`/`Seeded<&T>`; this brings with it better error messages, less boilerplate and extensibility
 - **added:** `Sample4` and `Sample4a` helper trait
 - **added:** `Perlin`, `Value`, `CellValue`, `CellDistance`, `CellDistanceSq` now support 4D sampling
 - **added:** `tileable` method and `Tileable` to create a 2D tileable noise from a 4D noise